Almost working version.
I printed this one with Cura's 'Super Quality' profile, (0.12 layer
height, 20% infill -- not that there is much to infill), no supports.

It works, but is a bit loose around the battery; plus because I still
can't get a working push-pin, I had to use nuts and bolts, and there
isn't clearance for the nut.

Plus the Charge lamp is obscured.  Added a 'right' option that puts
the battery case on the right hand side; this obscures the Power LED
instead.

Signed-off-by: Peter Chubb <peter@chubb.wattle.id.au>